feat: 상태별 태스크 개수 조회 TODO, IN_PROGRESS, DONE 테스트 코드 작성 Resolves: #14
